refactor: move thread runtime event store owner
